requirements.txt
Django
djangorestframework
pylint-django
psycopg2-binary
python-dotenv
django-cors-headers
